backend/pyproject.toml
[project]
name = "costcourter-backend"
version = "0.1.0"
description = "FastAPI backend for CostCourter."
authors = [{name = "CostCourter"}]
requires-python = ">=3.11"
dependencies = [
    "fastapi>=0.115.0",
    "uvicorn[standard]>=0.30.0",
    "sqlmodel>=0.0.22",
    "sqlalchemy>=2.0.32",
    "alembic>=1.13.2",
    "psycopg[binary]>=3.2.1",
    "pydantic-settings>=2.4.0",
    "python-multipart>=0.0.9",
    "python-jose[cryptography]>=3.3.0",
    "celery>=5.4.0",
    "redis>=5.0.7",
    "structlog>=24.1.0",
    "httpx>=0.27.0",
    "typer>=0.12.3",
    "orjson>=3.10.3",
    "email-validator>=2.1.0",
    "pymysql>=1.1.0",
    "croniter>=2.0.3",
    "beautifulsoup4>=4.14.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.3.2",
    "pytest-asyncio>=0.23.8",
    "pytest-cov>=5.0.0",
    "black>=24.4.2",
    "ruff>=0.5.6",
    "mypy>=1.10.0",
    "marimo>=0.16.0",
    "pandas>=2.3.2",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.ruff]
line-length = 100
exclude = ["alembic", "notebooks"]

[tool.ruff.lint]
select = ["E", "F", "I", "N", "UP", "B", "C4", "TID", "ISC"]
ignore = ["E203", "E266", "E501", "B008"]

[tool.pytest.ini_options]
asyncio_mode = "auto"
pythonpath = ["app"]

[tool.hatch.metadata]
allow-direct-references = true

[tool.hatch.build.targets.wheel]
packages = ["app"]

[tool.mypy]
python_version = "3.11"
disallow_untyped_defs = true
warn_return_any = true
warn_unused_configs = true
warn_unused_ignores = true

[[tool.mypy.overrides]]
module = ["celery", "celery.*"]
ignore_missing_imports = true

[[tool.mypy.overrides]]
module = ["jose", "jose.*"]
ignore_missing_imports = true

[[tool.mypy.overrides]]
module = ["croniter", "croniter.*"]
ignore_missing_imports = true

[[tool.mypy.overrides]]
module = ["apprise", "apprise.*"]
ignore_missing_imports = true

[[tool.mypy.overrides]]
module = ["pandas", "pandas.*"]
ignore_missing_imports = true

[tool.ruff.lint.per-file-ignores]
"app/models/*.py" = ["UP037"]

[dependency-groups]
dev = [
    "pytest>=8.3.2",
    "pytest-asyncio>=0.23.8",
    "pytest-cov>=5.0.0",
    "black>=24.4.2",
    "ruff>=0.5.6",
    "mypy>=1.10.0",
    "marimo>=0.16.0",
    "pandas>=2.3.2",
]
